+++ client/ecs/GraphicSystem.hpp
@@ -38,7 +38,7 @@ namespace ECS {
              * @param SceneType
              * @param deltaTime
              */
-            void update(SceneManager &sceneManager, int deltaTime, std::vector<Network::Packet> &packetQueue, Network::PacketManager &packetManager) override;
+            void update(SceneManager &sceneManager, float deltaTime, std::vector<Network::Packet> &packetQueue, Network::PacketManager &packetManager) override;
 
         private:
             sf::Event _event;
