+++ ecs/components/PositionComponent.hpp
@@ -23,7 +23,7 @@ namespace ECS {
              * @param y
              * @param uid
              */
-            PositionComponent(int x, int y);
+            PositionComponent(float x, float y);
             /**
              * @brief Destroy the position Component object
              *
@@ -49,7 +49,7 @@ namespace ECS {
              * @param x
              * @param y
              */
-            void move(int x, int y);
+            void move(float x, float y);
 
             std::shared_ptr<IComponent> clone() const override;
         protected:
@@ -58,12 +58,12 @@ namespace ECS {
              * @brief define the x position of the entity
              *
              */
-            int _x;
+            float _x;
             /**
              * @brief define the y position of the entity
              *
              */
-            int _y;
+            float _y;
     };
 }
